The _info module is in google.cloud.
