+++ plugins/modules/gcpubsub_facts.py
@@ -1,0 +1,1 @@
+cloud/google/gcpubsub_facts.py
